src/main/java/com/softserve/careactivities/utils/dateformat/DateMapper.java
package com.softserve.careactivities.utils.dateformat;

import org.springframework.stereotype.Component;

import java.time.LocalDateTime;
import java.time.ZoneId;
import java.time.ZonedDateTime;

@Component
public class DateMapper {

    public ZonedDateTime asZonedDateTime(LocalDateTime localDateTime) {
        return localDateTime != null ? ZonedDateTime.of(localDateTime, ZoneId.systemDefault()) : null;
    }

    public LocalDateTime asLocalDateTime(ZonedDateTime zonedDateTime) {
        return zonedDateTime != null ? zonedDateTime.toLocalDateTime() : null;
    }

}
